use response_block, even when expects mis-matches
